--- main.c
@@ -1,8 +1,0 @@
-#include "function.h"
-
-int main() {
-    system("title ½ò¦a¹p  v3.0.0 ");
-    system("mode con cols=120 lines=36");
-    game_dif_choose();
-    return 0;
-}
